feat: remove Credential, move createPresentation to AttestedClaim
