+++ packages/actors_api/src/actors/Claimer.ts
@@ -18,16 +18,15 @@ import type {
   IRequestForAttestation,
 } from '@kiltprotocol/types'
 import Message from '@kiltprotocol/messaging'
-import Credential from '../credential/Credential'
 
 /**
- * Creates a presentation for an arbitrary amount of [[Credential]]s which can be verified in [[verifyPresentation]].
+ * Creates a presentation for an arbitrary amount of [[AttestedClaim]]s which can be verified in [[verifyPresentation]].
  *
- * @param identity The Claimer [[Identity]] which owns the [[Credential]]s.
+ * @param identity The Claimer [[Identity]] which owns the [[AttestedClaim]]s.
  * @param message The message which represents multiple [[CType]]s, [[IRequestClaimsForCTypes]]s and whether privacy
  * enhancement is supported.
  * @param verifier The [[IPublicIdentity]] of the verifier that requested the presentation.
- * @param credentials The [[Credential]]s which should be verified.
+ * @param credentials The [[AttestedClaim]]s which should be verified.
  * @throws [[ERROR_PE_MISMATCH]], [[ERROR_MESSAGE_TYPE]], [[ERROR_PE_CREDENTIAL_MISSING]].
  * @returns A message which represents either an array of [[AttestedClaim]]s if privacy enhancement is not supported
  * or a combined presentation. Both of these options can be verified.
@@ -36,7 +35,7 @@ export function createPresentation(
   identity: Identity,
   message: IMessage,
   verifier: IPublicIdentity,
-  credentials: Credential[]
+  credentials: AttestedClaim[]
 ): Message {
   // did we get the right message type?
   if (message.body.type !== Message.BodyType.REQUEST_CLAIMS_FOR_CTYPES) {
@@ -95,7 +94,7 @@ export type ClaimerAttestationSession = {
  * include into the [[Attestation]] as legitimations.
  * @param option.delegationId The unique identifier of the desired delegation.
  * @returns A message containing an [[IRequestAttestationForClaim]] and a [[ClaimerAttestationSession]] which together with an [[ISubmitAttestationForClaim]]
- * object can be used to create a [[Credential]].
+ * object can be used to create an [[AttestedClaim]].
  */
 export function requestAttestation(
   claim: IClaim,
@@ -132,23 +131,23 @@ export function requestAttestation(
 }
 
 /**
- * Builds a [[Credential]] which can be verified when used in [[createPresentation]].
+ * Builds a [[AttestedClaim]] which can be verified when used in [[createPresentation]].
  *
  * @param message The session object corresponding to the [[ISubmitAttestationForClaim]].
  * @param session The [[ClaimerAttestationSession]] which corresponds to the message and [[AttestedClaim]].
- * @returns A signed and valid [[Credential]].
+ * @returns A signed and valid [[AttestedClaim]].
  */
 export function buildCredential(
   message: IMessage,
   session: ClaimerAttestationSession
-): Credential {
+): AttestedClaim {
   if (message.body.type !== Message.BodyType.SUBMIT_ATTESTATION_FOR_CLAIM) {
     throw SDKErrors.ERROR_MESSAGE_TYPE(
       message.body.type,
       Message.BodyType.SUBMIT_ATTESTATION_FOR_CLAIM
     )
   }
-  return Credential.fromRequestAndAttestation(
+  return AttestedClaim.fromRequestAndAttestation(
     session.requestForAttestation,
     message.body.content.attestation
   )
