+++ packages/actors_api/src/actors/Verifier.spec.ts
@@ -2,13 +2,12 @@
  * @group unit/actor
  */
 
-import { CType, Identity } from '@kiltprotocol/core'
+import { AttestedClaim, CType, Identity } from '@kiltprotocol/core'
 import type { IClaim, ICType } from '@kiltprotocol/types'
 import { mockChainQueryReturn } from '@kiltprotocol/chain-helpers/lib/blockchainApiConnection/__mocks__/BlockchainQuery'
 import Message from '@kiltprotocol/messaging'
 import { Crypto } from '@kiltprotocol/utils'
 import { Attester, Claimer, Verifier } from '..'
-import Credential from '../credential/Credential'
 
 jest.mock(
   '@kiltprotocol/chain-helpers/lib/blockchainApiConnection/BlockchainApiConnection'
@@ -20,7 +19,7 @@ describe('Verifier', () => {
   let verifier: Identity
   let cType: CType
   let claim: IClaim
-  let credential: Credential
+  let credential: AttestedClaim
   const blockchainApi = require('@kiltprotocol/chain-helpers/lib/blockchainApiConnection/BlockchainApiConnection')
     .__mocked_api
