+++ packages/actors_api/src/index.ts
@@ -1,7 +1,6 @@
 import * as Attester from './actors/Attester'
 import * as Claimer from './actors/Claimer'
 import * as Verifier from './actors/Verifier'
-import Credential from './credential/Credential'
 import type * as types from './types'
 
-export { Verifier, Claimer, Attester, Credential, types }
+export { Verifier, Claimer, Attester, types }
