+++ packages/actors_api/src/types.ts
@@ -3,7 +3,6 @@
  * @module ActorsTypes
  */
 
-export { ICredential } from './credential/Credential'
 export {
   IPresentationReq,
   IPartialRequest,
