+++ packages/sdk-js/src/index.ts
@@ -1,12 +1,7 @@
 import * as core from '@kiltprotocol/core'
 import * as Actors from '@kiltprotocol/actors-api'
 import Message, * as Messaging from '@kiltprotocol/messaging'
-import {
-  Credential,
-  Claimer,
-  Attester,
-  Verifier,
-} from '@kiltprotocol/actors-api'
+import { Claimer, Attester, Verifier } from '@kiltprotocol/actors-api'
 import { BlockchainUtils } from '@kiltprotocol/chain-helpers'
 import * as ChainHelpers from '@kiltprotocol/chain-helpers'
 import * as Utils from '@kiltprotocol/utils'
@@ -17,7 +12,6 @@ export {
   Message,
   Messaging,
   Actors,
-  Credential,
   Claimer,
   Attester,
   Verifier,
@@ -31,7 +25,6 @@ export default {
   Message,
   Messaging,
   Actors,
-  Credential,
   Claimer,
   Attester,
   Verifier,
